+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create wallets with 12 words mnemonics by default
 
 ### Fixed
+- Handle long messages in system message dialog
 
 ## [0.1.7] - 2021-05-21
 ### Added